extra detail for the previous commit:
So there was a problem in the git with push and commit so I couldn't save the previous commit.
The previous commit: styled the home page.
The home page has 10 buttons that each of them is for a project.
Now the project1 is compeleted and the other buttons are disable.
And the profile page has 4 social media link.
Also we can switch to these pages without any refreshing and the projects on the homw page are the same.